1st draft at passing SSH user into auth HTTP API
