+++ lib/auth/tun.go
@@ -21,6 +21,7 @@ import (
 	"fmt"
 	"io"
 	"net"
+	"net/http"
 	"os"
 	"sort"
 	"sync"
@@ -51,10 +52,8 @@ const dialRetryInterval = time.Duration(time.Millisecond * 50)
 type AuthTunnel struct {
 	// authServer implements the "beef" of the Auth service
 	authServer *AuthServer
-	// apiServer maintains a map of API servers, each assigned
-	// to a certain role. this allows to break Auth API into
-	// a set of restricted API services with well-defined permissions
-	apiServer *APIWithRoles
+	config     *APIConfig
+
 	// sshServer implements the nuts & bolts of serving an SSH connection
 	// to create a tunnel
 	sshServer       *sshutils.Server
@@ -106,13 +105,13 @@ func SetLimiter(limiter *limiter.Limiter) ServerOption {
 // a tunnel server and having tun clients connect.
 func NewTunnel(addr utils.NetAddr,
 	hostSigners []ssh.Signer,
-	apiServer *APIWithRoles,
+	apiConf *APIConfig,
 	authServer *AuthServer,
 	opts ...ServerOption) (tunnel *AuthTunnel, err error) {
 
 	tunnel = &AuthTunnel{
 		authServer: authServer,
-		apiServer:  apiServer,
+		config:     apiConf,
 	}
 	tunnel.limiter, err = limiter.NewLimiter(limiter.LimiterConfig{})
 	if err != nil {
@@ -322,6 +321,7 @@ func (s *AuthTunnel) handleWebAgentRequest(sconn *ssh.ServerConn, ch ssh.Channel
 
 // handleDirectTCPIPRequest accepts an incoming SSH connection via TCP/IP and forwards
 // it to the local auth server which listens on local UNIX pipe
+//
 func (s *AuthTunnel) handleDirectTCPIPRequest(sconn *ssh.ServerConn, sshChannel ssh.Channel, req *sshutils.DirectTCPIPReq) {
 	defer sconn.Close()
 
@@ -332,12 +332,38 @@ func (s *AuthTunnel) handleDirectTCPIPRequest(sconn *ssh.ServerConn, sshChannel 
 		return
 	}
 
-	// Forward this new SSH channel to API-with-roles server. It will try to proxy this
-	// connection to the API service mapped to this role:
-	if err := s.apiServer.HandleNewChannel(sconn.RemoteAddr(), sshChannel, role); err != nil {
-		log.Error(err)
-		return
-	}
+	api := NewAPIServer(&AuthWithRoles{
+		authServer:  s.config.AuthServer,
+		permChecker: s.config.PermissionChecker,
+		sessions:    s.config.SessionService,
+		role:        role,
+		alog:        s.config.AuditLog,
+	})
+	fs := makefakeSocket()
+
+	go func() {
+		connection := &FakeSSHConnection{
+			remoteAddr: sconn.RemoteAddr(),
+			sshChan:    sshChannel,
+			closed:     make(chan int),
+		}
+		select {
+		// Accept() will unblock this select
+		case fs.connections <- connection:
+		}
+		// wait for the connection to close:
+		select {
+		case <-connection.closed:
+			fs.Close()
+		}
+	}()
+
+	// serve HTTP API via this SSH connection until it gets closed:
+	http.Serve(&fs, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		log.Infof("[API] %v %v", r.Method, r.URL.String())
+		r.Header.Set("TELEPORT_USER", sconn.User())
+		api.ServeHTTP(w, r)
+	}))
 }
 
 func (s *AuthTunnel) keyAuth(
